requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
attrs==25.3.0
bcrypt==4.3.0
bitarray==3.3.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.8
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
greenlet==3.2.1
hexbytes==1.3.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
packaging==25.0
parsimonious==0.10.0
pluggy==1.5.0
propcache==0.3.1
psycopg2-binary==2.9.10
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyotp==2.9.0
pytest==8.3.5
pytest-flask==1.3.0
python-dotenv==1.1.0
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
SQLAlchemy==2.0.40
toolz==1.0.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
web3==7.10.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.0
